requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==19.3.1
greenlet==2.0.2
iniconfig==2.0.0
inquirerpy==0.3.4
mysql-connector-python==8.1.0
packaging==23.1
pfzy==0.3.4
pluggy==1.3.0
prompt-toolkit==3.0.39
protobuf==4.21.12
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.20
tomli==2.0.1
typing_extensions==4.7.1
wcwidth==0.2.6
